audio_upload_server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
yt-dlp==2024.11.18
mutagen==1.47.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==11.0.0
